pyproject.toml
[tool.poetry]
name = "my_fastapi_project"
version = "0.1.0"
description = "A FastAPI application for processing CSV files"
authors = ["dibyendubar13702@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
pandas = "^2.0.2"
requests = "^2.31.0"
sqlalchemy = "^2.0.11"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"  

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
